+++ client/src/main/components/datatable/Datatable.jsx
@@ -9,7 +9,7 @@ import axios from "axios";
 
 const Datatable = ({ column }) => {
   const location = useLocation();
-  const path = location.pathname.split("/")[2];
+  const path = location.pathname.split("/")[1];
   console.log(path)
   const [list, setList] = useState([]);
   const { data } = useFetch(`/${path}`)
@@ -25,7 +25,6 @@ const Datatable = ({ column }) => {
 
       await axios.delete(`http://localhost:5500/api/${path}/${id}`, { withCredentials: false }
       );
-      // await axios.delete(`https://stay-solutions.herokuapp.com/api/${path}/${id}`);
 
       setList(list.filter((item) => item._id !== id));
     } catch (err) {
