+++ client/src/main/components/navbar/navbar.scss
@@ -46,6 +46,7 @@
         box-shadow: 2px 4px 10px 1px rgba(201, 201, 201, 0.47);
         border-radius: 10px;
         display: none;
+        z-index: 1000;
 
         li {
           width: 180px;
