+++ client/src/main/datatablesource.js
@@ -21,3 +21,17 @@ export const taskColumns = [
         width: 100,
     },
 ];
+
+export const updateColumns = [
+    { field: "_id", headerName: "ID", width: 100 },
+    {
+        field: "title",
+        headerName: "Update",
+        width: 150,
+    },
+    {
+        field: "desc",
+        headerName: "Description",
+        width: 500,
+    },
+];
